added InventoryController and moved inventory methods from ItemsController to it
